moves css and logo to public folder and changes relevant paths so app doesnt break
