+++ src/index.js
@@ -3,7 +3,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 
-import './index.css';
+import '../public/index.css';
 
 ReactDOM.render(
   <App />,
